Live in tmux
When opening a new shell when tmux is not running then run tmux new

If tmux is already running, give a choice to connect to an existing
session or create a new one (or qQ to not).

With these dotfiles it is not recommended to work outside of tmux due to
https://github.com/neovim/neovim/issues/2048.  While there are many
solutions in the issue thread the one that works best for me is to
rebind c-h in tmux to send the proper command to neovim.  YMMV.